+++ src/commonMain/kotlin/de/tsenger/vdstools/vds/tr03171/StatusIndicator.kt
@@ -1,7 +1,0 @@
-package de.tsenger.vdstools.vds.tr03171
-
-enum class StatusIndicator {
-    NONE,
-    BLOCKLISTING,
-    ALLOWLISTING
-}